requirements.txt
requests>=2.32.3
openai>=1.0.0
python-dotenv>=1.0.1
